AM-1107: refactored consumer pact for create and roles
