+++ pacts/am_org_role_mapping-am_role_assignment_service_roles.json
@@ -1,0 +1,56 @@
+{
+  "provider": {
+    "name": "am_role_assignment_service_roles"
+  },
+  "consumer": {
+    "name": "am_org_role_mapping"
+  },
+  "interactions": [
+    {
+      "description": "RAS takes s2s/auth token and returns list of roles",
+      "request": {
+        "method": "GET",
+        "path": "/am/role-assignments/roles"
+      },
+      "response": {
+        "status": 200,
+        "headers": {
+          "Content-Type": "application/vnd.uk.gov.hmcts.role-assignment-service.get-roles+json;charset=UTF-8;version=1.0"
+        },
+        "body": [
+          {
+            "name": "judge",
+            "label": "Judge - Sample role (Only for Testing)",
+            "description": "Judicial office holder able to do judicial case work",
+            "category": "JUDICIAL"
+          },
+          {
+            "name": "tribunal-caseworker",
+            "label": "Tribunal Caseworker",
+            "description": "Tribunal caseworker",
+            "category": "STAFF"
+          },
+          {
+            "name": "senior-tribunal-caseworker",
+            "label": "Senior Tribunal Caseworker",
+            "description": "Senior Tribunal caseworker",
+            "category": "STAFF"
+          }
+        ]
+      },
+      "providerStates": [
+        {
+          "name": "A list of roles are available in role assignment service"
+        }
+      ]
+    }
+  ],
+  "metadata": {
+    "pactSpecification": {
+      "version": "3.0.0"
+    },
+    "pact-jvm": {
+      "version": "4.1.7"
+    }
+  }
+}
